wrap implementation
