+++ lib/application.js
@@ -9,7 +9,7 @@ const isGeneratorFunction = require('is-generator-function');
 const debug = require('debug')('koa:application');
 const onFinished = require('on-finished');
 const response = require('./response');
-const compose = require('koa-compose');
+const compose = require('../../compose');
 const isJSON = require('koa-is-json');
 const context = require('./context');
 const request = require('./request');
@@ -99,14 +99,11 @@ module.exports = class Application extends Emitter {
    * Old-style middleware will be converted.
    *
    * @param {Function} fn
-   * @param {Object} opts
    * @return {Application} self
    * @api public
    */
 
-  use(fn, opts) {
-    opts = opts || { wrappers: this.wrappers };
-
+  use(fn) {
     if (typeof fn !== 'function') throw new TypeError('middleware must be a function!');
     if (isGeneratorFunction(fn)) {
       deprecate('Support for generators will been removed in v3. ' +
@@ -115,11 +112,21 @@ module.exports = class Application extends Emitter {
       fn = convert(fn);
     }
 
-    if (opts.wrappers && Array.isArray(opts.wrappers))
-      fn._wrappers = opts.wrappers;
-
     debug('use %s', fn._name || fn.name || '-');
     this.middleware.push(fn);
+    if (this.wrappers && this.wrappers.length) this.wrap();
+    return this;
+  }
+
+  wrap(wrappers) {
+    wrappers = wrappers || this.wrappers;
+    if (!wrappers || !this.wrappers.length) return this;
+    this.middleware[this.middleware.length - 1]._wrappers = wrappers;
+    return this;
+  }
+
+  dewrap() {
+    delete this.middleware[this.middleware.length - 1]._wrappers;
     return this;
   }
